separate parser, calculating add calculate_satpos
