+++ satpos.py
@@ -1,234 +1,79 @@
 
-from io import TextIOWrapper
 import math
+import parse_rinex
 
 
-GM = 3.986005 * pow(10, 14)
-OMEGA_e = 7.292115 * pow(10, -5)
+GM = 3.986005 * 10**14
+OMEGA_e = 7.292115 * 10**(-5)
 
+def _calculate_tk(t, toe) -> float:
+    tk = t - toe
+    if tk > 302400.0:
+        tk = tk - 604800.0
+    elif tk < -302400.0:
+        tk = tk + 604800.0
 
-class EndOfFile(Exception):
-    pass
+    return tk
 
-class ErrorOBSRecord(Exception):
-    pass
+def _calculate_Ek(Mk, e):
+    Ek = Mk
+    temp = Ek
+    while math.fabs(Ek-temp) >= 1e-10:
+        temp = Ek
+        Ek = Mk + e*math.sin(Ek)
 
-def _split_neg_num(number, start_index=0):
-    index_minus = number.find('-', start_index)
-    fixed = []
+    return Ek
 
-    if index_minus > 0 and not number[index_minus-1].isalpha():
-        num1 = number[:index_minus]
-        num2 = number[index_minus:]
-        fixn1 = _split_neg_num(num1)
-        fixn2 = _split_neg_num(num2)
-        
-        if fixn1 is None:
-            fixed.append(num1)
-        else:
-            for i in fixn1:
-                fixed.append(i)
-
-        if fixn2 is None:
-            fixed.append(num2)
-        else:
-            for i in fixn2:
-                fixed.append(i)
-
-        return fixed
-
-    else:    
-        if index_minus != -1:
-            return _split_neg_num(number, index_minus+1)
-        else:
-            return None
-
-
-def _fix_negative_num(nums: list) -> list:
-    fixed_nums = []
-    for num in nums:
-        fixed_num = _split_neg_num(num)
-        if fixed_num is not None:
-            for fn in fixed_num:
-                fixed_nums.append(fn)
-        else:
-            fixed_nums.append(num)
-
-    return fixed_nums
-
-
-def _skip_header(rinex_file: TextIOWrapper) -> None:
-    header_end_founded = False
-
-    while True:
-        line = rinex_file.readline()
-        if 'END' not in line and not header_end_founded:
-            continue
-        elif 'END' in line:
-            header_end_founded = True
-            break
-
-
-def _read_PRN_EPOCH_SV_CLK(nums: list) -> dict:
-    if len(nums) != 10:
-        raise ErrorOBSRecord(str(nums))
+def calculate_satpos(sat_rinex: dict):
+    A = sat_rinex['sqrt_A']**2
+    n_0 = math.sqrt(GM/ (A**3))
+    n = n_0 + sat_rinex['Delta_n']
+    e = sat_rinex['e_Eccentricity']
+    tk = _calculate_tk(sat_rinex['TTM'],sat_rinex['Toe'])
+    Mk = sat_rinex['M0'] + n * tk
+    Ek = _calculate_Ek(Mk, e)
     
-    return {
-        'PRN': nums[0],
-        'EPOCH': {
-            'YEAR': nums[1],
-            'MONTH': nums[2],
-            'DAY': nums[3],
-            'HOUR': nums[4],
-            'MINUTE': nums[5],
-            'SECOND': nums[6],
-        },
-        'SV_clock_bias':float(nums[7]),
-        'SV_clock_drift':float(nums[8]),
-        'SV_clock_drift_rate':float(nums[9])
-    }
-    
+    vk = math.atan2(math.sqrt(1-e*e) * math.sin(Ek), math.cos(Ek) - e)
+    phi_k = vk + sat_rinex['omega']
 
-def _read_BROADCAST_ORBIT_1(nums: list) -> dict:
-    if len(nums) != 4:
-        raise ErrorOBSRecord(str(nums))
+    d_uk = sat_rinex['Cuc'] * math.cos(2*phi_k) + sat_rinex['Cus'] * math.sin(2*phi_k)
+    d_rk = sat_rinex['Crc'] * math.cos(2*phi_k) + sat_rinex['Crs'] * math.sin(2*phi_k)
+    d_ik = sat_rinex['Cic'] * math.cos(2*phi_k) + sat_rinex['Cis'] * math.sin(2*phi_k)
 
-    return {
-        'IODE': float(nums[0]),
-        'Crs': float(nums[1]),
-        'Delta_n': float(nums[2]),
-        'M0': float(nums[3])
-    }
+    uk = phi_k + d_uk
+    rk = A * (1 - e * math.cos(Ek)) + d_rk
+    ik = sat_rinex['i0'] + d_ik + sat_rinex['IDOT'] * tk
 
-def _read_BROADCAST_ORBIT_2(nums: list) -> dict:
-    if len(nums) != 4:
-        raise ErrorOBSRecord(str(nums))
+    xk_prim = rk * math.cos(uk)
+    yk_prim = rk * math.sin(uk)
 
-    return {
-        'Cuc': float(nums[0]),
-        'e_Eccentricity': float(nums[1]),
-        'Cus': float(nums[2]),
-        'sqrt_A': float(nums[3])
-    }
+    omega_k = sat_rinex['OMEGA'] + (sat_rinex['OMEGA_DOT'] - OMEGA_e) * tk - OMEGA_e * sat_rinex['Toe'] 
 
-def _read_BROADCAST_ORBIT_3(nums: list) -> dict:
-    if len(nums) != 4:
-        raise ErrorOBSRecord(str(nums))
+    xk = xk_prim * math.cos(omega_k) - yk_prim * math.cos(ik) * math.sin(omega_k)
+    yk = xk_prim * math.sin(omega_k) - yk_prim * math.cos(ik) * math.cos(omega_k)
+    zk = yk_prim * math.sin(ik)
 
-    return {
-        'Toe': float(nums[0]),
-        'Cic': float(nums[1]),
-        'OMEGA': float(nums[2]),
-        'CIS': float(nums[3])
-    }
-
-def _read_BROADCAST_ORBIT_4(nums: list) -> dict:
-    if len(nums) != 4:
-        raise ErrorOBSRecord(str(nums))
-
-    return {
-        'i0': float(nums[0]),
-        'Crc': float(nums[1]),
-        'omega': float(nums[2]),
-        'OMEGA_DOT': float(nums[3])
-    }
-
-def _read_BROADCAST_ORBIT_5(nums: list) -> dict:
-    if len(nums) != 4:
-        raise ErrorOBSRecord(str(nums))
-
-    return {
-        'IDOT': float(nums[0]),
-        'Codes_L2_channel': float(nums[1]),
-        'GPS_week': float(nums[2]),
-        'L2_P': float(nums[3])
-    }
-
-def _read_BROADCAST_ORBIT_6(nums: list) -> dict:
-    if len(nums) != 4:
-        raise ErrorOBSRecord(str(nums))
-
-    return {
-        'SV_accuracy': float(nums[0]),
-        'SV_health': float(nums[1]),
-        'TGD': float(nums[2]),
-        'IODC': float(nums[3])
-    }
-
-def _read_BROADCAST_ORBIT_7(nums: list) -> dict:
-    if len(nums) != 2:
-        raise ErrorOBSRecord(str(nums))
-
-    return {
-        'TTM': float(nums[0]),
-        'Fit_interval': float(nums[1])
-    }
-
-
-def _next_line(rinex_file: TextIOWrapper) -> list:
-    line = rinex_file.readline()
-    if line == '':
-        raise EndOfFile
-
-    nums = [num for num in line.strip().replace('D', 'e').split(' ') if num != '']
-    fixed_nums = _fix_negative_num(nums)
-
-    return fixed_nums
-
-
-def _extract_data(rinex_file: TextIOWrapper) -> dict:    
-    ext_data = {}
-    while True:
-        try:
-            str_data = []
-
-            for _ in range(8):
-                str_data.append(_next_line(rinex_file))                
-
-            
-            ex_data_l1 = _read_PRN_EPOCH_SV_CLK(str_data[0])
-            
-            ext_data[str(ex_data_l1['PRN'])] = ex_data_l1
-            ext_data[str(ex_data_l1['PRN'])].update(_read_BROADCAST_ORBIT_1(str_data[1]))
-            ext_data[str(ex_data_l1['PRN'])].update(_read_BROADCAST_ORBIT_2(str_data[2]))            
-            ext_data[str(ex_data_l1['PRN'])].update(_read_BROADCAST_ORBIT_3(str_data[3]))            
-            ext_data[str(ex_data_l1['PRN'])].update(_read_BROADCAST_ORBIT_4(str_data[4]))            
-            ext_data[str(ex_data_l1['PRN'])].update(_read_BROADCAST_ORBIT_5(str_data[5]))
-            ext_data[str(ex_data_l1['PRN'])].update(_read_BROADCAST_ORBIT_6(str_data[6]))            
-            ext_data[str(ex_data_l1['PRN'])].update(_read_BROADCAST_ORBIT_7(str_data[7]))
-
-        except EndOfFile:
-            print("End of file")
-            break
-        except ErrorOBSRecord as eobsr:
-            print('Error OBS Record', eobsr)
-            #skip records of the satellite
-    return ext_data
+    return (xk, yk, zk)
 
 
 def calculate_positions(sat_data: dict) -> dict:
     sat_pos = {}
-    for key, sat_rinex in sat_data.items():
-        sat_pos[key] = {}
 
-        A = math.pow(sat_rinex['sqrt_A'], 2)
-        n_0 = math.sqrt(GM/ math.pow(A, 3))
-        T = 2.0 * math.pi / n_0
-        n = n_0 + sat_rinex['Delta_n']
-        t_k = 0 # Need calculate !!!!!!
-        M_k = sat_rinex['M0'] + n * t_k
-        E_k = M_k 
-        
+    for _, sat_rinex in sat_data.items():
+        xk, yk, zk = calculate_satpos(sat_rinex)
 
+        sat_pos[str(sat_rinex['PRN'])] = {
+            'x': xk,
+            'y': yk,
+            'z': zk
+        }
+    
+    return sat_pos
 
-def read_rinex(filename: str):
-    ext_data = None
-    with open(filename, 'r') as rinex_file:
-        _skip_header(rinex_file)
-        ext_data = _extract_data(rinex_file)
-    return ext_data
 
 if __name__ == '__main__':
-    sat_data = read_rinex('./test/gmez0380.21n')
-    pos_sats = calculate_positions(sat_data)
-    print(pos_sats)
+    sat_data = parse_rinex.read_rinex('./test/940779338I_1.17N')
+    print(len(sat_data))
+    for key, rinex_info in sat_data.items():
+        pos = calculate_satpos(rinex_info)
+        print(key, pos)
